Abbreviate the action digest used in build directory paths
We currently add the full SHA sum of the action to the build path. The
downside of doing this is that the pathname may become quite long.
Though this is typically not a problem due to PATH_MAX being relatively
high, it does become annoying if you want to create UNIX sockets inside
build directories. Those have a limit that is a lot lower
(sizeof(sockaddr_un::sun_path)).