Let BatchedStoreBlobAccess cache errors of previous flushes
StorageFlushingBuildExecutor is currently implemented in such a way that
it may return ExecuteResponse messages of objects that did not actually
get uploaded into the CAS. This is problematic, as it may cause Bazel to
attempt to access files in the CAS that are nonexistent.

This change alters BatchedStoreBlobAccess in such a way that it will
hold on to an error that can be read and discarded by flushing. This
prevents StorageFlushingBuildExecutor from returning the original
response that contains the potentially missing objects.