Introducing a new Remote Worker protocol
The existing 'scheduler' protocol that bb_worker and bb_scheduler use to
communicate, has a number of serious flaws:

- Short hiccups in network connections cause build actions to fail
  unnecessarily.
- Workers have no identity, meaning the scheduler can't return any
  meaningful status.
- Workers cannot announce any platform properties, meaning a single
  scheduler cannot be used to handle multiple types of work.

Let's go ahead and add a new protocol that fixes the issues above. Just
like the old one, let's do our best to keep this protocol relatively
simple. Use a single Synchronize() RPC that workers call against the
scheduler. The worker reports its current state. In response, the
scheduler returns the state to which it desires the worker switches.