Let every build action have its own temporary directory
As an alternative to enabling cleaning of temporary directories, it may
be of interest to simply remove write access from it entirely. We could
let bb_runner override the TMPDIR environment variable for all build
actions. This variable could point to a directly placed in the build
directory. This directly will be purged automatically when the action
completes.

The advantage of this approach is that it's a lot more granular than
having simple temporary directory cleaning. There is a lower chance of
accidental sharing of data.