Allow bb_worker to manage multiple build directories
In the old days, bb_worker would manage a single build directory and
would call into a single bb_runner process to spawn build actions. Later
on this was extended, so that multiple bb_runner processes may be used.
This allows every build action to run in its own container, potentially
having its own resource limits.

This commit changes the cardinality once again: bb_worker may now manage
one or more build directories, each having one or more runners
associated with it. The reason for this change is that makes it easier
to achieve full file system level isolation between build actions. This
is a requirement for being able to spin up multi-tenant build clusters.

At the same time, this change renames 'local build directory' to 'native
build directory'. The motivation behind this change is that the
directory doesn't necessarily need to be local. It's perfectly fine to
use an NFS mount if you really wanted to. The important thing is that
it's a native POSIX directory that's being used.

This commit changes the schema of the bb_worker configuration. A
configuration that looks like this:

  {
    localBuildDirectory: { /* build directory options go here */ },
    runners: [ /* runners go here */ ],
  }

Now needs to be translated to this:

  {
    buildDirectories: [{
      native: { /* build directory options go here */ },
      runners: [ /* runners go here */ ],
    }],
  }