InMemoryBuildQueue: Add a primitive for safely terminating workers
This change adds a MarkTerminatingAndWait() function to
InMemoryBuildQueue. It can be used to indicate that workers are about to
be terminated. This causes a temporary drain to be installed on them.

MarkTerminatingAndWait() blocks until all matching workers have returned
back to the idle state, or have disappeared. This allows the caller to
go on and safely shut down the workers.