Improve robustness of HardlinkingFileFetcher
HardlinkingFileFetcher can end up getting stuck when the files in the
cache directory no longer match up with the in-memory bookkeeping. Let
it deal with that more gracefully by properly handling ENOENT/EEXIST
errors.

While there, improve the error messages returned by this type. This
makes it easier to figure out what's going on in the future.