Let bb-scheduler take the Bazel invocation ID into account
Right now bb-scheduler places all of the operations for a given platform
into a single queue and processes them based on (priority, queue
timestamp). This has a couple of disadvantages:

- Operations with a higher priority effectively fully dominate the build
  queue. They are preferred over lower priority operations
  unconditionally. This means that there is no safe way to say: "Execute
  this build a bit quicker than usual".

- Not all concurrent invocations of Bazel will get their fair share of
  the cluster. In fact, there will most likely be some oscillating,
  where one client is able to do lots of work, followed by being
  starved, followed by being able to do lots of work again.

- In-flight deduplication doesn't cause priorities to be raised. This
  means that a high-priority build may get slowed down in case one of
  its operations gets deduplicated against a low-priority one.

Let's tackle all of this by making the layout of objects tracked by the
scheduler a bit more advanced. It used to be this:

- InMemoryBuildQueue
  - platformQueue
    - operation

This change turns this into:

- InMemoryBuildQueue
  - platformQueue
    - invocation
      - operation
        - task

The operation struct has been decomposed into operation & task. Tasks
are what needs to be scheduled on workers, while operations are what
Bazel sees on the client side. This decomposition allows us to create
multiple operations for one task, thereby making it possible for a task
to have multiple priorities. When the scheduler decides to schedule a
task, all of its associated operations are removed from the scheduling
queue at once.

The additional invocation struct between platformQueue and operation
allows us to keep track of separate queues of operations for every
invocation of Bazel. The selection process of operations then becomes a
two-step process: we pick the most preferred invocation, followed by the
most preferred queued operation inside that invocation.

The second step (operation selection) remains identical to before: we
pick the operation with the highest priority, followed by the lowest
queueing timestamp. The first step (invocation selection) is the most
interesting part now: we also take the current number of executing
operations for an invocation into account. The higher the number of
executing operations is, the less favourable the invocation is. This
means that the scheduler will automatically balance towards giving every
invocation its fair share.

When implemented naïvely, this means priorities would be ignored between
invocations entirely; it would only take effect inside of a single
build. This is of course not what we want. We still want to provide a
facility for prioritizing certain builds over others. To solve that,
multiply the number of executing operations with a slowly growing
exponential function. This makes it possible to prioritize builds in a
fine-grained manner.

The base of the exponential function has been chosen to be
2^0.01 =~ 1.007. This means that a build that has a priority value that
is 100 lower than another one, the scheduler will attempt to give it
twice as many workers.