Add a checker for path writability to bb_runner
It's often the case that bb_runner is spawned in environments where one
or more global directories are writable (e.g., /tmp). Administrators of
such clusters might want to add some checks that this list of writable
paths doesn't grow over time.

This change adds a configuration option to let bb_runner scan the file
system on startup. Whenever a file or directory is found that is
writable, but not listed in bb_runner's configuration file, it will fail
to start.

Written by: Daniel Wagner-Hall