Let bb_worker deal with bb_runner disappearing gracefully
This change extends BuildClient to detect cases where build actions
complete in a catastrophic way. This may happen in all sorts of cases
(e.g., storage failures), but one important case is when the runner
process is unavailable. When this happens, we set a new flag in the
Synchronize() request: prefer_being_idle.

When set, the scheduler will allow the worker to enter the idle state,
regardless of whether more work is available. When the worker is idle,
it can perform more readiness checks against the runner process. The
worker can thus safely suspend until the runner process comes back
online again.

THe changes to the remote worker protocol will also make it possible to
implement graceful worker termination in case of signals. This isn't
done as part of this PR, but can easily be added later on.

Fixes: #33