Port bb-remote-execution to the new filesystem API
Filenames are now stored in the new path.Component type. This change
adjusts all of the code to make use of it. The FUSE file system also
uses it now, meaning that it's impossible for invalid filenames to end
up in directories.

Furthermore, all pathname processing has been rewritten to use
path.Resolve(). This means that we properly support "." and ".." and can
differentiate between absolute and relative paths.