Add PrepopulatedDirectory.FilterChildren()
Inside the client-side FUSE daemon we want to add logic to automatically
remove files from the file system that are no longer present remotely.
We'll solve this by scanning the entire bazel-out/ directory at the
start of every build to gather the digests. For every batch of files,
we'll call FindMissingBlobs(), thereby causing all of those files to be
touched.

The files that are present will then be guaranteed to remain in
centralized storage for the duration of the build. The ones that are
absent can be removed.