Add NativeLeaf.GetOutputServiceFileStatus()
The Remote Output Service API provides a BatchStat() function that can
be used to obtain the status of a large collection of files at once,
including the REv2 digest of those files. This commit adds a method to
all NativeLeaf objects, allowing us to extract the status of single
files. This function will be used by bb_clientd to implement
BatchStat().