bb_scheduler: Replace the argv[0] column with the Bazel target name
Now that REv2's RequestMetadata has been extended to include the Bazel
target name, let's adjust the scheduler UI to show this instead of
argv[0]. The advantage of showing the target name is that can be
obtained without loading the Command message from the CAS, thereby
reducing the load on the scheduler. Do add some compatibility code to
show argv[0] if the target name is not present.

At the same time, make sure we respect platform properties specified in
the Action message. REv2 states that both these fields need to be
considered.