Allow workers to compute (monetary) costs of build actions
This change adds a new decorator for BuildExecutor that inspects the
running time of a build action and uses that to determine how much money
it ended up costing. This information will get stored inside the
ExecutedActionMetadata's auxiliary_metadata, meaning that it gets stored
in the Action Cache.

The first step will be to extend bb_browser to show this information. In
the future we'll work towards adding facilities for logging this
information to some external service, so that aggregated costs of full
builds may be computed.