bb_clientd: Add a schema for persisting output paths
This change adds a Protobuf schema for how we can persist the contents
of individual output paths managed by bb_clientd to disk. Ideally we'd
place this inside of the bb_clientd repository, but that seems pretty
hard. The types in pkg/filesystem/fuse need to be extended to know how
to generate directory entries in this format.