Allow workers to provide their size to the scheduler
It's a common convention to put resource requirements of actions in
platform properties. In our case we want to add logic to bb_scheduler to
allow it to automatically pick the right worker based on outcomes of
previous executions. For that to work, we want multiple sizes of workers
to announce themselves using the same properties. The size class number
will then be used to distinguish them further.