Implement process table cleaning on macOS
This change adds a local patch to golang.org/x/sys, allowing us to issue
sysctl() calls to obtain the process table contents. The sysctl() call
will return kinfo_proc structs, for which we now have generated
bindings.

This approach is a lot faster than using some of the existing libraries
that are out there, as those tend to parse the output of ps(1).