Add fake_python
Bazel currently requires that remote execution environments ship with a
copy of Python. They need this copy of Python to launch py_binary()
targets, even if another copy of Python is present in the input root.
This is because the Python stub scripts that Bazel emits have a
'#!/usr/bin/env python' at the top.

This change adds a helper binary named fake_python that you can use as a
drop-in replacement for /usr/bin/python. It effectively parses the stub
scripts, extracting the full path of the Python stored in the input
root. It then relaunches the stub script using that copy instead.

The advantage of using this executable is that you can force people to
bring their own copies of Python, making it easier to ensure
compatibility in the long run.