Switch invocation ID to google.protobuf.Any
It's already a oneof type containing RequestMetadata or a message
indicating that background learning is performed. By changing it to
google.protobuf.Any, it becomes easier for people to develop their own
mechanisms for keying operations to which fairness should be applied.
For example, one may build its own implementation that takes the
credentials of the user into account.