FUSE: Compensate execution timeout with BlobAccess read times
When using FUSE-based workers, the input root is populated lazily. This
means that the 'downloading inputs' phase completes instantly. This does
mean that the execution itself may be slowed down by network I/O. This
approach may end up causing flakiness in case of unexpected network
latency. Tests that normally succeed may fail due to timeouts.

Let's fix this by making bb_worker compensate the execution timeout with
the time it took to read data through BlobAccess. This change adds a
new implementation of Clock called SuspendableClock. Any Context and
Timer created through it may be suspended/resumed by calling
Suspend()/Resume() on the SuspendableClock. We then provide a
SuspendingBlobAccess that decorates the existing storage objects,
calling into the SuspendableClock.

We also provide basic facilities for getting insight in the amount of
suspension on a per-action basis. We both expose this data as auxiliary
metadata and as a Prometheus histogram. We're also going to extend
bb-browser to display this information.