Add CompletedActionLogger
There are cases in which you may want to let Buildbarn generate a stream
of all of the actions that were built. Examples include:

- Being able to correlate events in Build Event Streams with activity on
  a build cluster.
- Being able to do cost accounting of actions that ran.

This change realizes exactly this. It extends bb_worker in such a way
that it can open a bidirectional streaming RPC to an external service.
HistoricalExecuteResponse messages are pushed through this stream; the
same as what bb_worker already stores into the CAS for all uncached
build results.

More details about this feature can be found in this ADR:

https://github.com/buildbarn/bb-adrs/blob/master/0006-operation-logging-and-monetary-resource-usage.md