Prevent a rare crash in PageRankStrategyCalculator
If an action has never run before, it will first get executed on the
largest worker, followed by the smallest. After completing execution on
the largest worker, an execution timeout for the smallest worker needs
to be computed. This is the point at which the crash may occur.

The crash is caused if due to rounding, the largest worker reported an
execution timeout in excess of the action's original timeout value. For
example, an action may have a 15 minute timeout. If the action finishes
at around the 15 minute mark, the worker may report it as 15 minutes and
a couple of milliseconds. This causes PageRankStrategyCalculator to
assume there are no samples at all, making it unable to compute a median
execution time.