Add ActionRouter
Before every an action gets queued by the scheduler, we currently need
to perform three different types of analysis:

1. We need to extract the platform, so that we know which platform queue
   to use.
2. We need to extract an invocation ID from the request metadata, so
   that we can place it in the right invocation context in the scheduler
   to do fair scheduling.
3. We need to do initial size class analysis on it, so that we pick a
   worker of the right size (in case multiple size classes are used).

All of these forms of analysis are currently performed independenly of
each other. I have noticed that this isn't flexible enough. Sometimes
you want to do remapping of actions between platforms, and in that case
the invocation ID extraction and initial size class analysis are also
affected.

This change introduces a new interface called ActionRouter, which
InMemoryBuildQueue uses to delegate all forms of pre-queueing analysis.
A simple implementation is provided that does all three forms of
analysis described above independently. Then there is also a
demultiplexing variant that allows you to configure the behaviour on a
per-platform basis.

I suspect that most users of Buildbarn that operate simpler setups
aren't really interested in configuring this extensively. This is why
it's recommended to use the following action router configuration as a
starting point:

    actionRouter: {
      simple: {
        platformKeyExtractor: { actionAndCommand: {} },
        invocationKeyExtractor: { requestMetadata: {} },
        initialSizeClassAnalyzer: {
          defaultExecutionTimeout: '1800s',
          maximumExecutionTimeout: '7200s',
        },
      },
    }

This example configuration is also documented in bb_scheduler.proto.