CompletedActionLogger: Add the ability to add instance name prefixes
If multiple build clusters push results to a single completed action
logging service, you likely want to add the cluster name to the
completed action messages generated by each of the individual clusters,
so that the originating cluster can be reobtained later on. To solve
this, we can use instance name prefixes.