Add ConvertCommandToShellScript()
This utility function can be used to convert an REv2 Command message to
a shell script that invokes the action in the way that's encoded in the
message. This is useful for running actions manually.

In order to implement this function, we want to be able to reuse
OutputHierarchy. This type currently depends on BuildDirectory, which is
a bit too broad. Let's add a new interface type,
ParentPopulatableDirectory, which contains just the methods of
BuildDirectory that OutputHierarchy uses.

This function will be used by bb_browser and bb_clientd going forward.