bb_scheduler: Allow grouping by correlated invocations ID
By passing in --build_request_id, multiple invocations of Bazel can set
the same UUID in the REv2 RequestMetadata message in a field called the
correlated invocations ID. This change makes it possible for
bb_scheduler to group actions by by this UUID, ensuring that fairness is
applied at that level.

To do this, we rename the existing 'requestMetadata' option to
'toolInvocationId'. This means that if you previously had the following
in your scheduler config:

    invocationKeyExtractors: [{ requestMetadata: {} }],

It should now become:

    invocationKeyExtractors: [{ toolInvocationID: {} }],

If you want to group actions both by correlated invocations ID and tool
invocation ID, you may use the following:

    invocationKeyExtractors: [
      { correlatedInvocationsId: {} },
      { toolInvocationId: {} },
    ],