Similar to fake_python, add fake_xcrun
If you want to deploy fake_python on a macOS system, you will need to
store it in a custom directory on the system and run 'xcode-select -s'
to repoint the development tools in /usr/bin/* to it. One requirement is
that this fake SDK also has a functional copy of xcrun.

This change brings in a stub implementation of xcrun that you can put in
place to make fake_python work.