Add ByteRangeLockSet
NFSv4 has integrated support for file locking. This means that if
processes issue calls to lockf(), flock(), or fcntl(), those will get
translated to LOCK, LOCKT or LOCKU operations. This requires us to
perform bookkeeping of which locks have been acquired.

It is common for NFSv4 servers to implement locking according to POSIX
semantics. POSIX locking semantics are fairly strong, in that every byte
in a file can have its own lock associated with it. There is also
support for both shared and exclusive locks.

This change adds a new generic data type, called a ByteRangeLock. In
addition to containing a range and lock type, it contains a comparable
field that acts as the owner. We need this field both to merge adjacent
locks, but also to perform queries.

ByteRangeLocks can be placed in a ByteRangeLockSet. These sets are
modeled as a linked list, sorted by starting offset. Every file managed
by the NFSv4 server will have its own ByteRangeLockSet.