Add an NFSv4 server to bb_worker
bb_worker has the ability to expose build directories through FUSE,
causing data to be loaded from the CAS on demand. This approach works
pretty well on Linux, but on macOS the user experience is far less
solid. One needs to install the OSXFUSE/macFUSE kernel extension to get
it working. There are some known issues with these kernel extensions,
causing them to lock up the system.

To sidestep this issue, this change adds an NFSv4 server to our
codebase, built on top of the same virtual file system layer as the FUSE
server. It uses the recently released go-xdr compiler to translate the
definitions in RFC 7531 to Go type definitions. This keeps the
implementation relatively small and (relatively) easy to understand and
test.

For now this implementation can only be used on macOS. It is in theory
also possible to use this on other systems, but it requires custom code
for establishing the mount. No effort has been spent to write this for
other systems yet.