Allow chmod() calls against CAS backed files
This was permitted in the past, but the VFS refactoring we did added
logic for rejecting it. This, unfortunately, causes build failures when
bb_clientd's Remote Output Service feature is used. Bazel seems to
generate calls to chmod() that aren't useful.

We should ideally patch up Bazel to no longer do this. That said, we do
need to remain compatible for the time being.