Generalize PreviousExecutionStatsStore
This data type is currently tailored towards the Initial Size Class
Cache (ISCC), but there is no need for it. The implementation is general
enough that it could also be used for the Action Cache (AC) or Indirect
Content Addressable Storage (ICAS).

This change renames this type to MutableProtoStore, with the BlobAccess
based implementation being renamed to BlobAccessMutableProtoStore.
Inside the scheduler code we add a typedef, both to keep the existing
code compiling and to be able to emit mocks.

A subtle difference is that BlobAccessPreviousExecutionStatsStore was
also responsible for initializing the Protobuf message upon load. This
can no longer be done, so we move that into FeedbackDrivenAnalyzer.