Let PoolBackedFileAllocator's VirtualClose() respect the count
Even though we extended VirtualClose() to take a count, we apparently
forgot to propagate that count into the f.release() call. For FUSE this
wasn't problematic, because there we only call VirtualClose(1). With
NFSv4 it does start to hurt, as it's not uncommon for the same file to
be opened multiple times and get deduplicated into a single state ID.