Add BuildExecutor.CheckReadiness()
Right now bb_worker only has logic for doing health checking against the
bb_runner process. There are no facilities for bb_worker to have
intrinsic health checks.

This change addresses this by giving BuildExecutor a method for doing
exactly that. We let the implementation of CheckReadiness() of
LocalBuildExecutor call into the runner, so that the existing behaviour
remains.

While there, we can move calls to CheckReadiness() into BuildClient, as
it can now access it properly.