Add TestInfrastructureFailureDetectingBuildExecutor
Bazel's test execution infrastructure provides a common way for tests to
report to the execution environment that an infrastructure failure has
occurred. This change adds a decorator for BuildExecutor that can make
use of this. It adds a simple implementation of BuildExecutor that
automatically causes subsequent readiness checks to fail if the number
of failing tests exceeds some kind of configured threshold.