Add AuthenticationMetadataKeyExtractor
Now that we have a way of displaying authentication metadata,
InMemoryBuildQueue can also use it for keying. This ensures fairness
between different users.

The main advantage of this change is that actions on the bb_scheduler
page are grouped by username when enabled, making it easier to find the
actions that belong to a given user.