NFSv4: Make READDIR work more reliably on macOS
macOS may sometimes issue READDIR requests that have the cookie verifier
set to zero, even though a previous response from the server returned a
concrete value for the client to use.

Even though it's in violation of the NFSv4, let's permit requests that
have the verifier explicitly set to zero.