Automatically unmount NFS mounts during shutdown
For FUSE it's not problematic to leave stale FUSE mounts around. For
NFSv4 it's a different story, as stale NFSv4 mounts cause processes to
hang when being accessed.

This change hooks into the existing terminationContext &
terminationGroup facilities that we already use for gracefully syncing
LocalBlobAccess, to call unix.Umount() upon process shutdown.