FUSE: Generalize the mechanism for setting Linux BDI options
Linux 6.2 has added sysfs attributes for toggling the
BDI_CAP_STRICTLIMIT flag on FUSE mounts:

https://github.com/torvalds/linux/commit/8e9d5ead865a1a7af74a444d2f00f1ef4539bfba
https://github.com/torvalds/linux/commit/27bbe9d48d4e298864e18b39f091342c68b81637

This means that it's finally possible to achieve decent performance on
FUSE when running build actions that use mmap() intensively, without
making local modifications to your system's kernel.

This change generalizes the mechanism we already had for setting the
max_ratio option, so that it's possible to configure arbitrary sysfs
attributes. This makes it possible to both set max_ratio and the newly
added strict_limit option.