NFSv4: Remove workarounds for older versions of macOS
The NFSv4 client that is part of macOS Ventura 13.3 includes a large
number of fixes that would either require workarounds on our end, or
would simply cause build actions to misbehave. I would therefore _very_
strongly recommend using this file system in combination with that
version of macOS or later.

Because of that, let's remove all of the compatibility workarounds for
earlier versions of macOS.