NFSv4/FUSE: Call unmount(2) repeatedly before mounting
Previous invocations of bb_worker and such may leave stale NFSv4/FUSE
mounts behind. This means that if bb_worker were to crashloop, these
need to be cleaned up. In the case of FUSE, I have observed that it may
also be permitted to place FUSE mounts inside of each other. Therefore,
let's add a utility function that calls in a loop until it fails.