bb_scheduler: Remove support for ASG lifecycle hooks
bb_scheduler supports reading ASG lifecycle hook events through AWS SQS,
allowing it to gracefully terminate workers running, for example, on
bare EC2 instances.

Even though this feature is useful to have, there is no longer any need
to have it integrated into the scheduler. The BuildQueueState API that
it uses internally to terminate the workers is nowadays reachable
through gRPC as well.

My plan is to add a new binary to the bb-autoscaler repository that can
be run as a separate deployment instead.