Let POSIXResourceUsage contain the termination signal name
REv2 uses a simple 'int32 exit_code' to capture the outcome of a
process. On POSIX-like systems, this is not sufficient. Processes can
also terminate due to signals, which is completely distinct from
terminating with an exit code. This differs from Windows, where any kind
of termination uses a simple integer exit code. If a process calls
exit() with the right integer value, the parent process may be tricked
into thinking it crashed due to a segfault.

If a process terminates due to a signal, os/exec will simply return exit
code -1, which isn't really helpful. To get the actual signal number
back, we need to extract it from the OS dependent
cmd.ProcessState.Sys().

This change extends the POSIXResourceUsage message that we attach to
every ActionResult to contain the signal that caused the process to
terminate. Let's make sure that LocalRunner on UNIX-like systems
populates it accordingly.

Signal numbers aren't really standardised by POSIX. SIGKILL doesn't
necessarily have to be equal to 9. Certain operating systems have custom
signals that aren't covered by POSIX (e.g., SIGINFO on BSD). Because of
that we store the signal in string form.

If the process terminated due to a signal, we still set the exit code to
some non-zero value. This is obviously required to keep remote execution
clients happy.