Make the handling of PATH compliant to recent versions of REv2 (#1724)
About 2 years ago there were discussions within the remote execution
working group on how argv[0] should be interpreted. The consensus was
that PATH lookups should be performed for paths that don't contain any
slashes. More details:

    https://github.com/bazelbuild/remote-apis/pull/210/files

It turns out that we already do this to a certain degree, but there are
certain flaws in how it's implemented:

- os/exec calls os.Getenv("PATH") to obtain the value of PATH, as
  opposed to using the value of the PATH environment that's set through
  cmd.Env.

- If PATH contains any relative paths, they are treated as being
  relative to the working directory of bb_runner. Not the value of
  cmd.Dir.

This change alters LocalRunner on UNIX-like systems to do their own PATH
resolution. By making sure that exec.CommandContext() is always called
with an absolute path, we can ensure that no flawed PATH lookups are
performed.