Catch up with changes to bb-storage/pkg/http
We now have a dedicated Protobuf configuration message for declaring
HTTP servers. Adjust bb_scheduler to use it properly.

While there, add an admin_route_prefix configuration option, so that the
scheduler web UI can be exposed on a subpath. This is necessary to make
OpenID Connect work properly. To support that we can't do any path
mangling.