Add authorization logic to the BuildQueueState service
Now that we have the ability to authenticate users visiting through the
bb_scheduler web UI, we can tackle this long-standing issue where
basically everyone is permitted to add/remove drains or kill operations
through the web UI. This action should be limited to cluster
administrators. Let's extend InMemoryBuildQueue to take two additional
authorizers for that.