Add metrics on the number of scheduler 'invocations' objects
With this change applied, you can run the following Prometheus query to
figure out how many 'invocations 'objects exist:

    buildbarn_builder_in_memory_build_queue_invocations_created_total
    -
    buildbarn_builder_in_memory_build_queue_invocations_removed_total

That also includes invocations that are idle. This means that there are
no queued/executing actions in them. They are only there because they
contain completed operations, or workers that most recently ran an
operation that belonged to it.

If you want to get metrics on the number of invocations that are active
(i.e., having either queued or executing actions in them), you may use
this query:

    buildbarn_builder_in_memory_build_queue_invocations_activated_total
    -
    buildbarn_builder_in_memory_build_queue_invocations_deactivated_total