Make OPEN_DOWNGRADE work with new VirtualClose() semantics
Right now we're using oofs.shareAccess to determine whether a file has
been opened redundantly. This is incorrect, as files may also be
downgraded using OPEN_DOWNGRADE. Even if oofs.shareAccess is reduced as
as a result of that, it may need to remain opened in its original form
for lock-owner files to work as expected.

We solve this by replacing the existing oofs.useCount field by a set of
counters for each of the share reservations (reading/writing) that are
supported. When one of these counters drop to zero, we can perform
(partial) close operations. These counters are stored in a new
shareCount helper type.